GOV-7671 Cherry-Pick If you fill in a List field in a workflow form and submit the form, the form doesn't get submitted. Workflow doesn't move to the next step

+++ d360.web/Models/Viewer Models.cs
@@ -428,7 +428,7 @@ namespace d360.web.Models
         public string TooltipContext { get; set; }
 
         [DataMember]
-        public int? TooltipID { get; set; }
+        public long? TooltipID { get; set; }
 
         [DataMember]
         public string TooltipUrl { get; set; }
